requirements.txt
pytorch=2.0.1
pytorch-cuda=11.7
pytorch-lightning=2.1.0
pytorch-mutex=1.0
numpy=1.26.0
pandas=2.0.3
tensorboard=2.15.0
tensorboard-data-server
